Change package 'model' to 'entity' and then ran mvn clean package
