+++ fifth-week-assignment/src/main/java/com/quintrix/java/training/iwterry/springboot/restapi/jpa/entity/Person.java
@@ -1,0 +1,58 @@
+package com.quintrix.java.training.iwterry.springboot.restapi.jpa.entity;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "person")
+public class Person {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private int id;
+  @Column(name = "first_name")
+  private String firstName;
+  @Column(name = "last_name")
+  private String lastName;
+  @Column(name = "age")
+  private int age;
+
+  public Person() {}
+
+  public Person(int id, String firstName, String lastName, int age) {
+    this.id = id;
+    this.firstName = firstName;
+    this.lastName = lastName;
+    this.age = age;
+  }
+
+  public int getId() {
+    return id;
+  }
+
+  public void setId(int id) {
+    this.id = id;
+  }
+
+  public String getFirstName() {
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public String getLastName() {
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+
+  public int getAge() {
+    return age;
+  }
+
+  public void setAge(int age) {
+    this.age = age;
+  }
+}
